backport m-c 1442648 - Fix mismatched types in mingw-w64 builds (const char16_t*/char16ptr_t).
When cross-compiling, fix mismatched types in nsMsgImapSearch.cpp. nsString::get()
returns a char16ptr_t (with MOZ_USE_CHAR16_WRAPPER) and that causes a mismatch so we cast it to a
const char16_t*. usAsciiCharSet is declared as a nsAutoString so we instead declare it as a char16_t* to
avoid subsequent type mismatches.